Remove the deprecated /admin view
Participant management, tournament name/description editing, and round
length editing are all already available in the normal setup/round flow
(and via MCP tools), making the separate admin page redundant. The one
non-admin-specific action it held, addParticipant, moves to
app/login/actions.ts where it's actually used (NewPlayerDrawer).
