+++ app/login/actions.ts
@@ -1,5 +1,6 @@
 "use server";
 import { cookies } from "next/headers";
+import { revalidatePath } from "next/cache";
 
 import { getSession } from "@/app/login/getSession";
 import {
@@ -7,6 +8,8 @@ import {
   createSessionToken,
   sessionCookieOptions,
 } from "@/app/login/session";
+import { room } from "@/app/constants";
+import { liveblocks } from "@/app/liveblocks/liveblocks";
 
 /**
  * Sets a signed session cookie that identifies the user for all the requests
@@ -28,6 +31,22 @@ export async function loginParticipant(username: string) {
   const cookieStore = await cookies();
   cookieStore.set(SESSION_COOKIE, token, sessionCookieOptions);
   console.info(`User ${username} logged in`);
+}
+
+export async function addParticipant(data: FormData) {
+  const username = (data.get("username") as string)?.trim();
+  if (!username) {
+    throw new Error("Username is required");
+  }
+  await liveblocks.mutateStorage(room, async ({ root }) => {
+    const participants = root.get("participants");
+    if (participants.findIndex((p) => p === username) !== -1) {
+      throw new Error("Username already exists");
+    }
+    participants.push(username);
+    console.log(`Added participant ${username}`);
+  });
+  revalidatePath("/login");
 }
 
 export async function logout() {
